modify pathBar and fileTable
